+++ algopy_testing/.vscode/settings.json
@@ -32,5 +32,6 @@
     "mypy.configFile": "pyproject.toml",
     // set to empty array to use config from project
     "mypy.targets": [],
-    "mypy.runUsingActiveInterpreter": true
+    "mypy.runUsingActiveInterpreter": true,
+    "cSpell.words": ["addw", "mulw"]
 }
